Trim PET PR to production files
