+++ package.json
@@ -12,7 +12,7 @@
     "reinstall": "pnpm run clean && pnpm install",
     "typecheck": "turbo run typecheck",
     "dev:web": "pnpm --filter @alook/web dev",
-    "dev:cli": "ALOOK_SERVER_URL=http://127.0.0.1:3003 ALOOK_PROJECT_ROOT=$PWD ALOOK_HEALTH_PORT=19515 pnpm --filter @alook/cli dev",
+    "dev:cli": "ALOOK_SERVER_URL=http://localhost:3000 ALOOK_PROJECT_ROOT=$PWD ALOOK_HEALTH_PORT=19515 pnpm --filter @alook/cli dev",
     "dev:email": "pnpm --filter @alook/email-worker dev",
     "dev:ws": "pnpm --filter @alook/ws-do dev",
     "dev:meeting": "node scripts/dev-meeting.mjs",
@@ -22,7 +22,6 @@
     "test:web": "pnpm --filter @alook/web test",
     "test:e2e": "pnpm --filter @alook/web test:e2e",
     "db:migrate": "pnpm --filter @alook/web exec wrangler d1 migrations apply alook-app --local",
-    "db:seed:homepage-stubs": "node scripts/seed-homepage-stubs.mjs",
     "db:reset": "rm -rf src/web/.wrangler/state && pnpm run db:migrate",
     "test:all": "vitest run",
     "deploy:web": "pnpm db:migrate:remote && cd src/web && npx opennextjs-cloudflare build && npx opennextjs-cloudflare deploy",
