+++ src/web/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.59",
   "private": true,
   "scripts": {
-    "dev": "next dev -H 127.0.0.1 -p 3003",
+    "dev": "next dev",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
