--- src/web/public/pets/tater/pet.json
@@ -1,6 +1,0 @@
-{
-  "id": "tater",
-  "displayName": "Tater",
-  "description": "A cheerful potato mascot with oversized translucent visor goggles, a glowing red stylus in one hand, soft pale limbs, and chunky futuristic sneakers, simplified into the Codex digital pet style.",
-  "spritesheetPath": "spritesheet.webp"
-}
