--- src/web/src/app/pet-preview/page.tsx
@@ -1,5 +1,0 @@
-import { PetPreviewPage } from "@/components/home-pet/pet-preview-page";
-
-export default function Page() {
-  return <PetPreviewPage />;
-}
